# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toonvill API",
        "description": "Toonvill API description",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://with42.kr/api",
            "description": "ToonVill Test Server"
        }
    ],
    "paths": {
        "/user/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "회원가입",
                "description": "회원가입",
                "operationId": "f832303921739b419e584927027fb638",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succeeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "로그인",
                "description": "로그인",
                "operationId": "485164a8acf9c7672983e88f841a6676",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succeeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "유효하지 않은 로그인 정보입니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "토큰 요청",
                "description": "로그인 이후 토큰 요청",
                "operationId": "531c6ff7dc00a74c728de11336f052a8",
                "responses": {
                    "200": {
                        "description": "응답성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "토큰 생성 성공"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "6d8a83ac57a3c392e1094335ff9eb2ad"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "example": "2022-09-21 23:55:52"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "유저 로그아웃",
                "description": "로그인 이후 로그아웃 요청",
                "operationId": "3eb379bb5078906212fa45f8bd7d6ac6",
                "responses": {
                    "200": {
                        "description": "응답성공"
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/me": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "유저 정보가져오기",
                "description": "유저 정보를 가져옵니다.",
                "operationId": "token",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "nickname": {
                        "type": "string",
                        "format": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "profile_url": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "login_method": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "kakao_token": {
                        "type": "string"
                    },
                    "naver_token": {
                        "type": "string"
                    },
                    "facebook_token": {
                        "type": "string"
                    },
                    "google_token": {
                        "type": "string"
                    },
                    "receive_push": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "integer"
                    },
                    "stop_start_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "stop_end_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "device_type": {
                        "type": "string"
                    },
                    "user_level": {
                        "type": "string"
                    },
                    "character_info": {
                        "type": "string",
                        "format": "json"
                    },
                    "remember_token": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Users": {
                "title": "Users",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "User--id": {
                "name": "User_id",
                "in": "path",
                "description": "Id of User",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": {}
                    }
                },
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    }
}